my-sdk/pyproject.toml
[build-system]
requires = ['hatchling']
build-backend = "hatchling.build"

[tool.hatch.version]
path = "my_sdk/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["my_sdk"]

[project]
name = 'my-sdk'
dynamic = ['version']
description = "My SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=2.7.0",
    "typing-extensions>=4.0.0"
]

[project.entry-points.apache_airflow_provider]
provider_info = "my_sdk:get_provider_info"